+++ src/screens/authorize/finds/PasswordResetScreen/PasswordResetScreen.style.ts
@@ -1,0 +1,3 @@
+import { StyleSheet } from 'react-native';
+
+const casswordResetScreenStyles = StyleSheet.create({});
